web-animations/animation-effect/animation-effect-clone.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>AnimationEffect clone() method test</title>
<meta name="assert" content="Creates and returns a new object of the same type as this object’s most-derived interface such that it will produce the same output as this object">
<link rel="help" href="http://w3c.github.io/web-animations/#dom-animationeffect-clone">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
test(function() {
    var effectsToTest = [newKeyframeEffect(), newMotionPathEffect()];
    effectsToTest.forEach(function(effect) {
        var cloned = effect.clone();

        assert_false(cloned === effect,
            'Cloned AnimationEffect object should be a copy of source object');
        assert_object_equals(cloned, effect, 'Cloned AnimationEffect object should be ' +
            'identical to source object');
    });
}, 'Test AnimationEffect.clone() creates a copy of AnimationEffect object ' +
    'with default options');

test(function() {
    var options = {
        name: 'someName',
        iterationComposite: 'accumulate',
        composite: 'add'
    };
    var effectsToTest = [newKeyframeEffect(options), newMotionPathEffect(options)];
    effectsToTest.forEach(function(effect) {
        var cloned = effect.clone();

        assert_false(cloned === effect,
            'Cloned AnimationEffect object should be a copy of source object');
        assert_object_equals(cloned, effect, 'Cloned AnimationEffect object should be ' +
            'identical to source object');
    });
}, 'Test AnimationEffect.clone() creates a copy of AnimationEffect object ' +
    'with not default options');
</script>
</body>
